Refactor: cambie la páginacion del dashboard, ya no se divide por roles
