--- frontend/pages/dashboard/administrator/addusers.tsx
@@ -1,15 +1,0 @@
-import Layoutp from "@/layouts/Layoutp";
-import styles from "@/styles/adminstrator-index.module.css";
-
-const AddUsers = () => {
-  return (
-    <Layoutp role="administrator">
-      <div>
-        <h1>Administrador</h1>
-        <h2>Añadir Usuarios</h2>
-      </div>
-    </Layoutp>
-  );
-};
-
-export default AddUsers;
