--- frontend/pages/dashboard/assistant/index.tsx
@@ -1,13 +1,0 @@
-import DragDrop from "@/components/DragDrop";
-import Layout from "@/layouts/Layout";
-import Layoutp from "@/layouts/Layoutp";
-
-const Assistant = () => {
-  return (
-    <Layoutp role="assistant">
-      <DragDrop />
-    </Layoutp>
-  );
-};
-
-export default Assistant;
