--- frontend/pages/dashboard/assistant/profile.tsx
@@ -1,12 +1,0 @@
-import { ConfigProfile } from "@/components/config-profile/ConfigProfile";
-import Layoutp from "@/layouts/Layoutp";
-
-const Profile = () => {
-  return (
-    <Layoutp role="assistant">
-      <ConfigProfile />
-    </Layoutp>
-  );
-};
-
-export default Profile;
