+++ frontend/pages/dashboard/list-users.tsx
@@ -1,17 +1,16 @@
 import FilterSection from "@/components/filter-section/FilterSection";
 import Table from "@/components/table/Table";
 import Layout from "@/layouts/Layout";
-import Layoutp from "@/layouts/Layoutp";
 
-const Administrator = () => {
+const ListUsers = () => {
   return (
-    <Layoutp role="administrator">
-      <div>
+    <Layout>
+      <>
         <FilterSection />
         <Table />
-      </div>
-    </Layoutp>
+      </>
+    </Layout>
   );
 };
 
-export default Administrator;
+export default ListUsers;
