--- frontend/pages/dashboard/professional/index.tsx
@@ -1,11 +1,0 @@
-import Layout from "@/layouts/Layout";
-
-const Professional = () => {
-  return (
-    <Layout>
-      <h1>Profesional Dashboard</h1>
-    </Layout>
-  );
-};
-
-export default Professional;
